requirements.txt
dash
dash-bootstrap-components
tensorflow
tensorflow-datasets
Pillow
rlds
numpy
pdash-draggable
